refactor(config): extract profile schema models from profile_loader
profile_loader.py (871 lines) mixed the typed profile schema dataclasses
(TradingConfig/StrategyConfig/RiskConfig/ExecutionConfig/SessionConfig/
MonitoringConfig/ProfileSchema + ProfileValidationError and session-time
parsing) with the YAML loading, override resolution, registry, and
BotConfig-building behavior. Split the schema into app/config/profile_schema.py,
leaving profile_loader to own loading/resolution.

profile_loader re-exports the schema types, so import sites are byte-stable with
identical class identity (verified, incl. from_yaml parsing). No behavior change.
